+++ MAServices/Interfaces/IReviewServices.cs
@@ -5,7 +5,7 @@ namespace MAServices.Interfaces
 {
     public interface IReviewServices
     {
-        Task PostNewReview(User user, Movie movie, string? descriptionVote, short vote, string? when);
+        Task PostNewReview(User user, Movie movie, string? descriptionVote, float vote, string? when);
 
         Task<List<ReviewDTO>?> SearchEngineReviews(User? user, Movie? movie);
     }
